# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  dto.AssignmentResponse:
    properties:
      id:
        example: 1
        type: integer
      quantity:
        example: 10
        type: integer
      resource_id:
        example: 1
        type: integer
      task_id:
        example: 1
        type: integer
    type: object
...
  dto.CreateAssignmentRequest:
    properties:
      quantity:
        example: 5
        type: integer
      resource_id:
        example: 1
        type: integer
      task_id:
        example: 1
        type: integer
    type: object
...
  dto.UpdateAssignmentRequest:
    properties:
      quantity:
        example: 10
        type: integer
      resource_id:
        example: 1
        type: integer
      task_id:
        example: 1
        type: integer
    type: object
...
  response.Response:
    properties:
      data: {}
      error:
        example: error message
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: shmonov.mv@gmail.com
    name: Shmonov Matvey
    url: https://t.me/Koshsky
  description: For managing the enterprise's universal resources
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Enterprise Resource Planning
  version: "1.0"
paths:
  /assignment:
    get:
      description: Get a list of all assignments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.AssignmentResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: List assignments
      tags:
      - Assignments
    post:
      consumes:
      - application/json
      description: Create a new assignment
      parameters:
      - description: Assignment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAssignmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AssignmentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create assignment
      tags:
      - Assignments
  /assignment/{id}:
    delete:
      description: Delete an assignment by ID
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Delete an assignment
      tags:
      - Assignments
    get:
      description: Get assignment by id
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AssignmentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get assignment
      tags:
      - Assignments
    put:
      consumes:
      - application/json
      description: Update an assignment by ID
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Assignment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAssignmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AssignmentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update an assignment
      tags:
      - Assignments
...
  /planning/processes:
    get:
      description: Get process planning
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get process planning
      tags:
      - Planning
  /planning/projects:
    get:
      description: Get project planning (project portfolio)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get project planning
      tags:
      - Planning
  /planning/tasks:
    get:
      description: Get task planning
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get task planning
      tags:
      - Planning
...
securityDefinitions:
  ApiKeyAuth:
    description: '"Provide JWT token in the format: Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"